[Paragraph] Move Next version to default
