+++ .storybook/stylesheets/kitten/_components.scss
@@ -110,4 +110,3 @@
 // Typography
 
 @include k-TypographyBodyCopy;
-@include k-Paragraph;
