+++ assets/javascripts/kitten/components/typography/paragraph/next.js
@@ -1,0 +1,1 @@
+export { Paragraph } from './'
