--- assets/javascripts/kitten/components/typography/paragraph/next/test.js
@@ -1,72 +1,0 @@
-import React from 'react'
-import renderer from 'react-test-renderer'
-import { Paragraph } from './index'
-
-describe('<Paragraph />', () => {
-  let component
-
-  describe('by default', () => {
-    beforeEach(() => {
-      component = renderer
-        .create(
-          <Paragraph
-            tag="p"
-            modifier="primary"
-            noMargin
-            normalLineHeight={false}
-            italic={false}
-            children="Lorem ipsum dolor sit amet…"
-          />,
-        )
-        .toJSON()
-    })
-
-    it('matches with snapshot', () => {
-      expect(component).toMatchSnapshot()
-    })
-  })
-
-  describe('with modifier', () => {
-    beforeEach(() => {
-      component = renderer
-        .create(<Paragraph tag="p" modifier="primary" />)
-        .toJSON()
-    })
-
-    it('matches with snapshot', () => {
-      expect(component).toMatchSnapshot()
-    })
-  })
-
-  describe('with no margin', () => {
-    beforeEach(() => {
-      component = renderer.create(<Paragraph tag="p" noMargin />).toJSON()
-    })
-
-    it('matches with snapshot', () => {
-      expect(component).toMatchSnapshot()
-    })
-  })
-
-  describe('without normal line height', () => {
-    beforeEach(() => {
-      component = renderer
-        .create(<Paragraph tag="p" normalLineHeight={false} />)
-        .toJSON()
-    })
-
-    it('matches with snapshot', () => {
-      expect(component).toMatchSnapshot()
-    })
-  })
-
-  describe('without italic', () => {
-    beforeEach(() => {
-      component = renderer.create(<Paragraph tag="p" italic={false} />).toJSON()
-    })
-
-    it('matches with snapshot', () => {
-      expect(component).toMatchSnapshot()
-    })
-  })
-})
