+++ assets/javascripts/kitten/next/index.js
@@ -1,5 +1,5 @@
 export { Modal } from '../components/modals/next'
-export { Stepper } from '../components/steppers/next/stepper'
+export { Stepper } from '../components/steppers/stepper' // Next version is now deprecated, please use normal version.
 export { Button } from '../components/buttons/button/button'
-export { Carousel } from '../components/carousel/carousel/carousel'
-export { Paragraph } from '../components/typography/paragraph/next'
+export { Carousel } from '../components/carousel/carousel' // Next version is now deprecated, please use normal version.
+export { Paragraph } from '../components/typography/paragraph' // Next version is now deprecated, please use normal version.
