+++ assets/stylesheets/kitten/_components.scss
@@ -71,4 +71,3 @@
 
 // Typography
 @import './components/typography/body-copy';
-@import './components/typography/paragraph';
